requirements.txt
# Core dependencies
python-dotenv>=1.0.0
colorama>=0.4.6
PyQt5>=5.15.9
pygame>=2.5.0
soundfile>=0.12.1
tiktoken>=0.4.0
opencv-python-headless>=4.7.0
numpy>=1.22.0
scipy>=1.11.2
torch>=2.1
transformers>=4.33.0
scikit-learn>=1.3.0
networkx>=3.0.0
groq>=0.9.0
keyboard>=0.13.5
simpleaudio>=1.0.4

# Audio processing
sounddevice>=0.4.6
pydub>=0.25.1
ffmpeg-python>=0.2.0
faster-whisper>=0.6.0

# Text-to-Speech
gtts>=2.2.3
parler-tts @ git+https://github.com/huggingface/parler-tts.git

# NLP and text processing
spacy>=3.5.0
textblob>=0.15.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz

# Visualization and graphics
PyOpenGL>=3.1.6
PyOpenGL_accelerate>=3.1.6
pybullet>=3.2.5
Pillow>=9.3.0
diffusers>=0.10.2  # Required for Stable Diffusion


# Additional utilities
datetime>=5.1
traceback2>=1.4.0

absl-py==2.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argbind==0.3.9
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
av==12.3.0
backoff==2.2.1
bcrypt==4.2.0
blis==0.7.11
build==1.2.1
CacheControl==0.14.0
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
cleo==2.1.0
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.6
confection==0.1.5
contourpy==1.2.1
crashtest==0.4.1
ctranslate2==4.3.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
DateTime==5.5
decorator==5.1.1
Deprecated==1.2.14
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
diffusers==0.30.0
distlib==0.3.8
distro==1.9.0
docstring_parser==0.16
dulwich==0.21.7
einops==0.8.0
en-core-web-sm==3.5.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.0
faster-whisper==1.0.3
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.15.4
fire==0.6.0
flatbuffers==24.3.25
flatten-dict==0.4.2
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
google-auth==2.33.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.9.0
grpcio==1.65.4
gTTS==2.5.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
installer==0.7.0
ipython==8.26.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
julius==0.2.7
keyboard==0.13.5
keyring==24.3.1
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.29
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.98
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
linecache2==1.0.0
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.2
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
opencv-python-headless==4.10.0.84
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
parso==0.8.4
pathlib_abc==0.1.1
pathy==0.11.0
pexpect==4.9.0
pillow==10.4.0
pipenv==2024.0.1
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
pooch==1.8.2
posthog==3.5.0
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==3.20.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybullet==3.2.6
pycparser==2.22
pydantic==1.10.17
pydantic_core==2.20.1
pydub==0.25.1
pygame==2.6.0
Pygments==2.18.0
pyloudnorm==0.1.1
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.1.2
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.4.1
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.2
randomname==0.2.1
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.7.1
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.4.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.32
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
textblob==0.18.0.post0
thinc==8.1.12
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.0
torch==2.4.0
torch-stoi==0.2.1
torchaudio==2.4.0
tqdm==4.66.5
traceback2==1.4.0
traitlets==5.14.3
transformers==4.43.3
trove-classifiers==2024.7.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
virtualenv==20.26.3
wasabi==1.1.3
watchfiles==0.23.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
zope.interface==7.0.1

